Changed searchCities in AccuWeatherService.kt to Observable stream, Added RecyclerView components, Added util for Model binding.

+++ app/src/main/java/com/gregrdm/accukipweatherme/api/model/TimeZone.kt
@@ -6,7 +6,7 @@ data class TimeZone(
     @SerializedName("Code")
     val code: String,
     @SerializedName("GmtOffset")
-    val gmtOffset: Int,
+    val gmtOffset: Float,
     @SerializedName("IsDaylightSaving")
     val isDaylightSaving: Boolean,
     @SerializedName("Name")
